Move ProxyOrderSyncer to OrderManagement engine
