+++ .rubocop_todo.yml
@@ -882,6 +882,7 @@ Style/FrozenStringLiteralComment:
     - 'engines/order_management/app/services/order_management/reports/enterprise_fee_summary/report_service.rb'
     - 'engines/order_management/app/services/order_management/reports/enterprise_fee_summary/scope.rb'
     - 'engines/order_management/app/services/order_management/reports/enterprise_fee_summary/summarizer.rb'
+    - 'engines/order_management/app/services/order_management/subscriptions/proxy_order_syncer.rb'
     - 'engines/order_management/app/services/order_management/subscriptions/subscription_summarizer.rb'
     - 'engines/order_management/app/services/order_management/subscriptions/subscription_summary.rb'
     - 'engines/order_management/app/services/reports.rb'
@@ -895,6 +896,7 @@ Style/FrozenStringLiteralComment:
     - 'engines/order_management/lib/order_management/engine.rb'
     - 'engines/order_management/lib/order_management/version.rb'
     - 'engines/order_management/order_management.gemspec'
+    - 'engines/order_management/spec/performance/order_management/subscriptions/proxy_order_syncer_spec.rb'
     - 'engines/order_management/spec/services/order_management/reports/enterprise_fee_summary/authorizer_spec.rb'
     - 'engines/order_management/spec/services/order_management/reports/enterprise_fee_summary/parameters_spec.rb'
     - 'engines/order_management/spec/services/order_management/reports/enterprise_fee_summary/permissions_spec.rb'
@@ -902,6 +904,7 @@ Style/FrozenStringLiteralComment:
     - 'engines/order_management/spec/services/order_management/reports/enterprise_fee_summary/renderers/html_renderer_spec.rb'
     - 'engines/order_management/spec/services/order_management/reports/enterprise_fee_summary/report_data/enterprise_fee_type_total_spec.rb'
     - 'engines/order_management/spec/services/order_management/reports/enterprise_fee_summary/report_service_spec.rb'
+    - 'engines/order_management/spec/services/order_management/subscriptions/proxy_order_syncer_spec.rb'
     - 'engines/order_management/spec/services/order_management/subscriptions/subscription_summizer_spec.rb'
     - 'engines/order_management/spec/services/order_management/subscriptions/subscription_summary_spec.rb'
     - 'engines/order_management/spec/spec_helper.rb'
@@ -954,7 +957,6 @@ Style/FrozenStringLiteralComment:
     - 'lib/open_food_network/products_and_inventory_report.rb'
     - 'lib/open_food_network/products_and_inventory_report_base.rb'
     - 'lib/open_food_network/property_merge.rb'
-    - 'lib/open_food_network/proxy_order_syncer.rb'
     - 'lib/open_food_network/rack_request_blocker.rb'
     - 'lib/open_food_network/referer_parser.rb'
     - 'lib/open_food_network/reports/bulk_coop_allocation_report.rb'
@@ -1212,7 +1214,6 @@ Style/FrozenStringLiteralComment:
     - 'spec/lib/open_food_network/permissions_spec.rb'
     - 'spec/lib/open_food_network/products_and_inventory_report_spec.rb'
     - 'spec/lib/open_food_network/property_merge_spec.rb'
-    - 'spec/lib/open_food_network/proxy_order_syncer_spec.rb'
     - 'spec/lib/open_food_network/referer_parser_spec.rb'
     - 'spec/lib/open_food_network/reports/report_spec.rb'
     - 'spec/lib/open_food_network/reports/row_spec.rb'
@@ -1306,7 +1307,6 @@ Style/FrozenStringLiteralComment:
     - 'spec/models/variant_override_spec.rb'
     - 'spec/performance/injection_helper_spec.rb'
     - 'spec/performance/orders_controller_spec.rb'
-    - 'spec/performance/proxy_order_syncer_spec.rb'
     - 'spec/performance/shop_controller_spec.rb'
     - 'spec/requests/checkout/failed_checkout_spec.rb'
     - 'spec/requests/checkout/paypal_spec.rb'
