+++ spec/controllers/admin/schedules_controller_spec.rb
@@ -90,7 +90,7 @@ describe Admin::SchedulesController, type: :controller do
 
         it "syncs proxy orders when order_cycle_ids change" do
           syncer_mock = double(:syncer)
-          allow(OpenFoodNetwork::ProxyOrderSyncer).to receive(:new) { syncer_mock }
+          allow(OrderManagement::Subscriptions::ProxyOrderSyncer).to receive(:new) { syncer_mock }
           expect(syncer_mock).to receive(:sync!).exactly(2).times
 
           spree_put :update, format: :json, id: coordinated_schedule.id, schedule: { order_cycle_ids: [coordinated_order_cycle.id, coordinated_order_cycle2.id] }
@@ -150,7 +150,7 @@ describe Admin::SchedulesController, type: :controller do
 
           it "sync proxy orders" do
             syncer_mock = double(:syncer)
-            allow(OpenFoodNetwork::ProxyOrderSyncer).to receive(:new) { syncer_mock }
+            allow(OrderManagement::Subscriptions::ProxyOrderSyncer).to receive(:new) { syncer_mock }
             expect(syncer_mock).to receive(:sync!).once
 
             create_schedule params
